+++ src/main/java/com/hotels/beeju/HiveMetaStoreJUnitRule.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2019 Expedia, Inc.
+ * Copyright (C) 2015-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -50,10 +50,24 @@ public class HiveMetaStoreJUnitRule extends BeejuJUnitRule {
    * Create a Hive Metastore with a pre-created database using the provided name and configuration.
    *
    * @param databaseName Database name.
-   * @param configuration Hive configuration properties.
+   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these with its defaults.
    */
-  public HiveMetaStoreJUnitRule(String databaseName, Map<String, String> configuration) {
-    super(databaseName, configuration);
+  public HiveMetaStoreJUnitRule(String databaseName, Map<String, String> preConfiguration) {
+    super(databaseName, preConfiguration);
+  }
+
+  /**
+   * Create a Hive Metastore with a pre-created database using the provided name and configuration.
+   *    
+   * @param databaseName Database name.
+   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these with its defaults.
+   * @param postConfiguration Hive configuration properties that will be set to override BeeJU's defaults.
+   */
+  public HiveMetaStoreJUnitRule(
+      String databaseName,
+      Map<String, String> preConfiguration,
+      Map<String, String> postConfiguration) {
+    super(databaseName, preConfiguration, postConfiguration);
   }
 
   @Override